Implemented the Account class.
The Account class serves as the foundation of our banking System.
It has Fields to identify a user and all of the banks features
can be accessed through his account. There is a "createAccountNo"
funtion which generates a 16 character long AccountNo which will act
as a primary key i.e used to identify a user. The class also has
value objects like Loan, Card and transactionHistory which aims to
add functionality for the user of the bank.