Implemented a transaction system.
The transaction system consists of a transaction class,
UtilLog class and a Constants Enumeration class.
This system allows for a user to track his transaction
details during his tenure as a user of this bank account.
The system maps a one-to-many relationship with a user's
account and his transactions. So a user can have multiple
transactions and all of those transactions are saved within
the transaction system. The transaction system is aided by an
enumeration object called Constants and a UtilLog object.

The enumeration object is used to categorize what kind of a
transaction it is. It is also used as a short descriptiono of
a transaction.

The UtilLog class aids the Transaction system by returning a
descriptive String about the transaction, how did it affected your
account and other things. It also uses the enumeration object to
return a valid description of a transaction.
In general it is used to get a log of whatever it is you are doing.