Created a BankingManagementSystem class and implemented its createAccount() Method.
We created a banking management System class which will automate the process of
creation of an account, given the required parameters. This class uses a CreateSession
Factory class to create an object of SessionFactory. This is done because creating a
session factory is a very resource intensive task which should be done only once per
application if you are using a single database. This is why a single class will be used
by all other classes to access that single session factory.

We also updated util log a bit to allow for users to create bank account at zero balance
as well. We also updated the creation of description to reflect those changes within the
transaction history.