Implemented the "updateAccount()" method for baking management system.
The updateAccount() method allows the user to update his account's
details on the fly. In order to update multiple fields of account,
the user needs to give a space separated list of options to the
console. The method then updates all those fields one by one as
the user gives the relevent information to the program.
The method uses an anuthenticateUser() method inside the Util class
to check the credibilty of the account given by the user and if the
account is in the database only then it proceeds further.
In order to avoid code repetition, we used a method called
"updateAccountDetails()" which generalises the  session.set%
process.
Finally we tested our implementation using Test.java.