Implemented the Loan class which will be the base for loan management system.
Here we have  loan class which will keep track of all the things
related to loans for a user. It is mapped to the user using a
one-to-one relation.

We also changed the datatype for the "balance" field in account
from long to double for more precision when we involve interest
rates from the bank into the balance of user.

configured the config.xml file accordingly and also tested the
loan class implemenatation using test class.