Implemented the "transferFunds()" method.
transferFunds allows you to transfer money either using cash
or from one's own account balance. On choosing cash, the money
is supposed to be submitted to the bank employee and he transfers
the funds as requested. On choosing account, the transferred fund
is deducted from users bank account itself. He is also given a choice
to transfer using cash in case he dosent have enough money.

We also increased the connection pool and created a Util method
called "twoOptionmenu()". As the name implies, this method is used
to create a text based two option menu which can take two values
and lets the user choose one of them. Option 1 is chosen if user
enters 1 else option 2 is chosen.
THERE IS NO WAY TO ABORT FROM THIS MENU.

We also updated UtilLog class to include a receiving constant
to describe the case when a user gets some amount transferred
to his account by someone else. enumeration RECEIVE is used for
it.

Finally we tested the method using the "Test.java" class.