Implemented the "withdrawOrDeposit"  function for the management system.
This method allows users to either withdraw from or deposit to his/her
bank account. In case of withdraw, the method  has an interesting
feature in that similar to an atm machine, it uses minimum coin
denominations to fullfill a withdrawal request. This is achieved
by using the coinExchange method whose implementation is based on
the popular dynamic programming (D.P) problem called the coin
exchange problem. This method however has some serious disadvantages.

1. The withdrawn amount must be an integer. If not the program itself
will use "floor" function to round it off to an integer. But only the rounded
off value is deducted from your account, and not the fraction.

2. The coinExhange method uses a 2d array. This matrix eats a lot
of memory, especially for large sums of money where a serious performance
hit will be seen. This is why currently it is able to support withdrawals
only upto  2^31-1 which is the maximu size of array in java.