Implemented the ATMCard class which form the base of atm management system.
The class has a lot of randomly generated fields which inclues the card
number, the mpin and the cvv. Out of the three only mpin can be changed.
As can be seen from the code the  the setters for a few of the fields are
private. This is done because as seen from reallife, the atms have some
information directly engraved within the card itself, which cannot be
changed without changing the card itself. This is why changing some
information is not allowed here as well. We can however allocate a completely
new card to user. The setter are still kept because hibernate uses them
to add information into proxy classes when we call get methods.

Finally, we have a 16 digit card no. these cards have fixed first
4 digits( ascii conversion of the string "Banking management system").
This is done to identify the bank.