Created an ATM machine class which will use atm cards to operate.
The atm machine uses atm cards to operate. It is an object meaning
it has different statuses and fields. The idea is to have multiple
atm machine objects stored within the database and when the user
asks for an atm machine, he could chose from a list of machines
based on their locations and whether they are active or not.

The ATM machine also has a  cashVault. This has a fixed amount of cash
at any given moment and  when the user withdraws cash from it, the cash
amount reduces. Finally at some point, someone has to fill the cashVault
again. That someone isnt yet decided.

The atm machine has functionalities which are almost the same as
the banking Management system. Hence it extends it.
It also adds features like pin change and paying emi's.
It can also ban a user if he uses up his 3 attempts for getting
his mpins right.

Finally we added a few enum constants and util logs to maintain
transaction logs for the added functionality of the atm.