Implemented a loan Management system class  for dealing with loans.
The loan management system class has two functions.

1. The takeLoan(), which returns a loan object which is then
attached to the users account. The formula for calculating
loan emi's  is tested and although it usually returns a
a double, a long value is used for easier payment of emi's.

2. The payLoanEmi() is used to pay the emi's if the user has
taken a loan. This gives you two options. Either pay using cash
which is to be either given to the bank employee or inserted within
the  atm machine, or deduct you emi from your bank coount balance
which fails if the user does not have enough balance.

3. The consequences of not paying the loan on time have not yet
been programmed.

Finally we also allowed the atm machine to use the method of the Loan
management system to  pay emis.