Added loan management to the banking management itself.
The bank management system is the centre for all the features
provided by this app. It essiatially aims to replicate actual banks
and hence has to  have all the functionalities.
The user can either pay his emi by going to the bank itself or
going to the nearest atm. But He can only take a loan by going to
the bank itself.