Implemented driver classes for the project.
Here we considered banks and atms as seperate entities and so
there are two different driver classes. One for the bankking management
system and the other for atm functionality.
These classes provide a text based input for users to interact with
the application. Both the banks and atms share functionality of
paying your loan emi's.

Looking at the code, it might seem like banks and atms share the same
codebase which is true, but atms can be played much more densely in an
area rather than a bank which means ease of access of banks
 functionalities  for the user.